More activities + buttons
This isn't all of the activities. Some of the buttons do things, some don't right now.  And we'll need to see if there are any errors. Also I can't tell if I need to give new IDs to all of the buttons--if so, I'll figure that out and do it in the next commit